pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Aleksey Kosin <akosin@udm.ru>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.8.1"
django = "^4.1.4"
dj-database-url = "^1.2.0"
gunicorn = "^20.1.0"
python-dotenv = "^0.21.0"
flake8 = "^6.0.0"
psycopg2-binary = "^2.9.5"
django-bootstrap4 = "^22.3"
whitenoise = "^6.2.0"
django-extensions = "^3.2.1"
ipython = "^8.7.0"
django-filter = "^22.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
